Added find ddo handler.
